pyproject.toml
[project]
name = "audio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "dotenv>=0.9.9",
    "eventlet>=0.40.4",
    "faster-whisper>=1.2.1",
    "flask>=3.1.2",
    "flask-socketio>=5.6.0",
    "numpy>=2.4.2",
    "openai>=2.21.0",
    "sounddevice>=0.5.5",
    "soundfile>=0.13.1",
    "tiktoken>=0.12.0",
    "webrtcvad-wheels>=2.0.14",
]
